+++ packages/transfer_manager_android/example/lib/main.dart
@@ -8,6 +8,7 @@ Future<void> main() async {
   final manager = TransferManager(
     engines: [
       AndroidBackgroundDownloadEngine(),
+      AndroidBackgroundUploadEngine(),
       TusTransferEngine(),
       HttpTransferEngine(),
     ],